Corrected collision to work with non-square window
Removed single Range check and replaced with a bounding rectangle,
and a renderArea rectangle. Collision now works by checking if the
renderArea is contained within the bounding rectangle. If not it
adjusts to place the renderArea at the bounding rectangle's relevent
edge and reverses the velocity.

Also updated config.ini with new values.